change to instant, no more delay
