+++ Server.API/Program.cs
@@ -115,12 +115,12 @@ using (var scope = app.Services.CreateScope())
         Cron.Daily(hour: 17) // fire at 00:00 Vietnam Time
                              //Cron.MinuteInterval(1) // Run every minute for testing
         );
-     recurringJobManager.AddOrUpdate<ITailoredReminderEmailService>(
-    "send-emergency-biometric-alert",
-        job => job.RunEmergencyBiometricJob(),
-        Cron.Minutely
-        //Cron.MinuteInterval(1) // Run every minute for testing
-    );
+    // recurringJobManager.AddOrUpdate<ITailoredReminderEmailService>(
+    //"send-emergency-biometric-alert",
+    //    job => job.RunEmergencyBiometricJob(),
+    //    Cron.Minutely
+    //    //Cron.MinuteInterval(1) // Run every minute for testing
+    //);
 
 }
