+++ Server.Application/HangfireService/TailoredReminderEmailService.cs
@@ -86,124 +86,124 @@ namespace Server.Application.HangfireService
             await _unitOfWork.SaveChangeAsync();
         }
 
-        public async Task SendEmergencyBiometricAlert(DateTime lastCheckTime)
-        {
-            var biometrics = await _unitOfWork.BasicBioMetricRepository.GetAllRecentBiometrics(lastCheckTime);
+        //public async Task SendEmergencyBiometricAlert(DateTime lastCheckTime)
+        //{
+        //    var biometrics = await _unitOfWork.BasicBioMetricRepository.GetAllRecentBiometrics(lastCheckTime);
 
-            foreach (var bio in biometrics)
-            {
-                var abnormalDetails = GetAbnormalitiesSummary(bio);
+        //    foreach (var bio in biometrics)
+        //    {
+        //        var abnormalDetails = GetAbnormalitiesSummary(bio);
 
-                if (abnormalDetails.Any())
-                {
-                    var email = bio.GrowthData?.GrowthDataCreatedBy?.Email;
-                    if (string.IsNullOrEmpty(email))
-                    {
-                        _logger.LogWarning($"Biometric {bio.Id}: email is null or empty.");
-                        continue;
-                    }
+        //        if (abnormalDetails.Any())
+        //        {
+        //            var email = bio.GrowthData?.GrowthDataCreatedBy?.Email;
+        //            if (string.IsNullOrEmpty(email))
+        //            {
+        //                _logger.LogWarning($"Biometric {bio.Id}: email is null or empty.");
+        //                continue;
+        //            }
 
-                    // Build a combined message
-                    var subject = "Emergency Biometric Alert";
-                    var message = "We detected the following abnormal readings:\n" +
-                                  string.Join("\n", abnormalDetails);
+        //            // Build a combined message
+        //            var subject = "Emergency Biometric Alert";
+        //            var message = "We detected the following abnormal readings:\n" +
+        //                          string.Join("\n", abnormalDetails);
 
-                    var reminder = new TailoredCheckupReminder
-                    {
-                        Id = Guid.NewGuid(),
-                        GrowthDataId = bio.GrowthDataId,
-                        Title = subject,
-                        Description = message,
-                        CheckupStatus = CheckupStatus.NotScheduled,
-                        Type = CheckupType.Emergency,
-                        CreatedBy = bio.GrowthData?.CreatedBy,
-                        CreationDate = _currentTime.GetCurrentTime(),
-                        IsActive = true
-                    };
+        //            var reminder = new TailoredCheckupReminder
+        //            {
+        //                Id = Guid.NewGuid(),
+        //                GrowthDataId = bio.GrowthDataId,
+        //                Title = subject,
+        //                Description = message,
+        //                CheckupStatus = CheckupStatus.NotScheduled,
+        //                Type = CheckupType.Emergency,
+        //                CreatedBy = bio.GrowthData?.CreatedBy,
+        //                CreationDate = _currentTime.GetCurrentTime(),
+        //                IsActive = true
+        //            };
 
-                    await _unitOfWork.TailoredCheckupReminderRepository.AddAsync(reminder);
-                    await _unitOfWork.SaveChangeAsync();
+        //            await _unitOfWork.TailoredCheckupReminderRepository.AddAsync(reminder);
+        //            await _unitOfWork.SaveChangeAsync();
 
-                    await _emailService.SendEmergencyBiometricAlert(
-                        email,
-                        subject,
-                        message
-                    );
+        //            await _emailService.SendEmergencyBiometricAlert(
+        //                email,
+        //                subject,
+        //                message
+        //            );
 
-                    _logger.LogInformation(
-                        $"[Hangfire] Created combined emergency reminder & sent email to {email} for biometric ID: {bio.Id}, reminder ID: {reminder.Id}"
-                    );
-                }
-            }
-        }
+        //            _logger.LogInformation(
+        //                $"[Hangfire] Created combined emergency reminder & sent email to {email} for biometric ID: {bio.Id}, reminder ID: {reminder.Id}"
+        //            );
+        //        }
+        //    }
+        //}
 
-        // what is abnormal readings, ref at diag lab
-        private List<string> GetAbnormalitiesSummary(BasicBioMetric bio)
-        {
-            var issues = new List<string>();
+        //// what is abnormal readings, ref at diag lab
+        //private List<string> GetAbnormalitiesSummary(BasicBioMetric bio)
+        //{
+        //    var issues = new List<string>();
 
-            // --- Blood Pressure ---
-            if (bio.SystolicBP.HasValue && bio.DiastolicBP.HasValue)
-            {
-                var sys = bio.SystolicBP.Value;
-                var dia = bio.DiastolicBP.Value;
+        //    // --- Blood Pressure ---
+        //    if (bio.SystolicBP.HasValue && bio.DiastolicBP.HasValue)
+        //    {
+        //        var sys = bio.SystolicBP.Value;
+        //        var dia = bio.DiastolicBP.Value;
 
-                if (sys >= 180 || dia >= 120)
-                    issues.Add($"Blood Pressure {sys}/{dia} mmHg which is Hypertensive Crisis (Seek emergency care)");
-                else if (sys >= 160 || dia >= 100)
-                    issues.Add($"Blood Pressure {sys}/{dia} mmHg which is Hypertension Stage 2");
-                else if (sys >= 140 || dia >= 90)
-                    issues.Add($"Blood Pressure {sys}/{dia} mmHg which is Hypertension Stage 1");
-                else if (sys < 90 || dia < 60)
-                    issues.Add($"Blood Pressure {sys}/{dia} mmHg which is Hypotension (Low Blood Pressure)");
-            }
+        //        if (sys >= 180 || dia >= 120)
+        //            issues.Add($"Blood Pressure {sys}/{dia} mmHg which is Hypertensive Crisis (Seek emergency care)");
+        //        else if (sys >= 160 || dia >= 100)
+        //            issues.Add($"Blood Pressure {sys}/{dia} mmHg which is Hypertension Stage 2");
+        //        else if (sys >= 140 || dia >= 90)
+        //            issues.Add($"Blood Pressure {sys}/{dia} mmHg which is Hypertension Stage 1");
+        //        else if (sys < 90 || dia < 60)
+        //            issues.Add($"Blood Pressure {sys}/{dia} mmHg which is Hypotension (Low Blood Pressure)");
+        //    }
 
-            // --- Blood Sugar (Gestational Diabetes references, fasting only) ---
-            if (bio.BloodSugarLevelMgDl.HasValue)
-            {
-                var sugar = bio.BloodSugarLevelMgDl.Value;
+        //    // --- Blood Sugar (Gestational Diabetes references, fasting only) ---
+        //    if (bio.BloodSugarLevelMgDl.HasValue)
+        //    {
+        //        var sugar = bio.BloodSugarLevelMgDl.Value;
 
-                // Always treat as fasting
-                if (sugar > 95)
-                    issues.Add($"Blood Sugar {sugar} mg/dL (when fasting) which is above pregnancy safe range (over 95 mg/dL) predicting a possible gestational diabetes");
-                else if (sugar < 70)
-                    issues.Add($"Blood Sugar {sugar} mg/dL (when fasting) which is Hypoglycemia risk");
-            }
+        //        // Always treat as fasting
+        //        if (sugar > 95)
+        //            issues.Add($"Blood Sugar {sugar} mg/dL (when fasting) which is above pregnancy safe range (over 95 mg/dL) predicting a possible gestational diabetes");
+        //        else if (sugar < 70)
+        //            issues.Add($"Blood Sugar {sugar} mg/dL (when fasting) which is Hypoglycemia risk");
+        //    }
 
-            // --- Heart Rate ---
-            if (bio.HeartRateBPM.HasValue)
-            {
-                var hr = bio.HeartRateBPM.Value;
-                if (hr > 120)
-                    issues.Add($"Heart Rate {hr} bpm which is Tachycardia (High Heart Rate)");
-                else if (hr < 50)
-                    issues.Add($"Heart Rate {hr} bpm which is Bradycardia (Low Heart Rate)");
-            }
+        //    // --- Heart Rate ---
+        //    if (bio.HeartRateBPM.HasValue)
+        //    {
+        //        var hr = bio.HeartRateBPM.Value;
+        //        if (hr > 120)
+        //            issues.Add($"Heart Rate {hr} bpm which is Tachycardia (High Heart Rate)");
+        //        else if (hr < 50)
+        //            issues.Add($"Heart Rate {hr} bpm which is Bradycardia (Low Heart Rate)");
+        //    }
 
-            // --- BMI ---
-            try
-            {
-                var bmi = bio.GetBMI();
-                if (bmi < 18.5f)
-                    issues.Add($"BMI {bmi:F1} which is Underweight");
-                else if (bmi > 35f)
-                    issues.Add($"BMI {bmi:F1} which is Obese");
-            }
-            catch (Exception ex)
-            {
-                _logger.LogWarning(ex, "BMI calculation failed for BasicBioMetric ID {BiometricId}", bio.Id);
-            }
+        //    // --- BMI ---
+        //    try
+        //    {
+        //        var bmi = bio.GetBMI();
+        //        if (bmi < 18.5f)
+        //            issues.Add($"BMI {bmi:F1} which is Underweight");
+        //        else if (bmi > 35f)
+        //            issues.Add($"BMI {bmi:F1} which is Obese");
+        //    }
+        //    catch (Exception ex)
+        //    {
+        //        _logger.LogWarning(ex, "BMI calculation failed for BasicBioMetric ID {BiometricId}", bio.Id);
+        //    }
 
-            return issues;
-        }
+        //    return issues;
+        //}
 
-        private static DateTime _lastCheckTime = DateTime.UtcNow.AddMinutes(-5);
+        //private static DateTime _lastCheckTime = DateTime.UtcNow.AddMinutes(-5);
 
-        public async Task RunEmergencyBiometricJob()
-        {
-            await SendEmergencyBiometricAlert(_lastCheckTime);
-            _lastCheckTime = DateTime.UtcNow;
-        }
+        //public async Task RunEmergencyBiometricJob()
+        //{
+        //    await SendEmergencyBiometricAlert(_lastCheckTime);
+        //    _lastCheckTime = DateTime.UtcNow;
+        //}
