+++ Server.Application/Services/BasicBioMetricService.cs
@@ -24,15 +24,17 @@ namespace Server.Application.Services
         private readonly IMapper _mapper;
         private readonly IClaimsService _claimsService;
         private readonly ICurrentTime _currentTime;
+        private readonly ITailoredCheckupReminderService _tailoredCheckupReminderService;
 
         public BasicBioMetricService(IUnitOfWork unitOfWork, IMapper mapper, IBasicBioMetricRepository basicBioMetricRepository,
-            IClaimsService claimsService, ICurrentTime currentTime)
+            IClaimsService claimsService, ICurrentTime currentTime, ITailoredCheckupReminderService tailoredCheckupReminderService)
         {
             _basicBioMetricRepository = basicBioMetricRepository;
             _unitOfWork = unitOfWork;
             _mapper = mapper;
             _claimsService = claimsService;
             _currentTime = currentTime;
+            _tailoredCheckupReminderService = tailoredCheckupReminderService;
         }
         public async Task<Result<List<ViewBasicBioMetricDTO>>> ViewAllBasicBioMetrics()
         {
@@ -157,6 +159,11 @@ namespace Server.Application.Services
             }
 
             var result = await _unitOfWork.SaveChangeAsync();
+
+            if (result > 0)
+            {
+                await _tailoredCheckupReminderService.SendEmergencyBiometricAlert(bbm.Id);
+            }
 
             return new Result<object>
             {
